[dv/alert_agent] async reset fix
Previously reset is issued with the clocking block. So if async mode is
on and one of the clock goes very slow, then there is a corner case
where all the reset of the signals are reset, but the slow clock's
interface is not.
To fix this issue, reset signals needs to be async (not driving through
the clocking block). To allow this, the input of the alert_esc_interface
is set to logic instead of wire.

Signed-off-by: Cindy Chen <chencindy@google.com>