[otbn] Run the ISS as a "stepped" subprocess in otbn_model
This patch doesn't actually change the behaviour of the code at all,
but should allow us to do cleverer things in future. It runs the
stepped simulator (instead of otbnsim) but, instead of running it with
system(..), it spawns a subprocess and communicates over stdin/stdout.

With future patches, we'll be able to actually run the ISS one cycle
at a time. Detection of when the ISS finishes isn't particularly
pretty (see the regex code in ISSWrapper::saw_busy_cleared), but it
should work for now. I'm assuming we'll completely overhaul the
tracing format in the reasonably near future anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>