[dv/gpio] support interrupt as a flop
According to PR #3422, now interrupt will take one more clock cycle to
be reset or reset. This affects GPIO scb prediction for interrupts. The
ideal solution would be to update the prediction logic for
`intr_state_update_queue`, but it seems to be quite complex. This PR
provides an easier fix to use the original `intr_state_update_queue`
calculated value, but wait one clock cycle before comparing with the
interrupt pin.

Signed-off-by: Cindy Chen <chencindy@google.com>