[doc] Remove the term "RFC" from the D2 checklist
The term RFC in OpenTitan context typically means an RFC which is
discussed within the Technical Committee. In the context of the
development lifecycle checklist, there shouldn't be the need for an RFC
handled by the TC when features are added between D1 and D2 stages.
(There *is* TC involvement when transitioning to the final development
stage. That's unchanged.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>