[kmac] Revise Assumption of process
There was a corner case that can violate the assumption but which is
legitimate input. When msg_valid_i and process_i are asserted at the
sametime, and if the message is at the end of the block, the state
machine doesn't send the keccak_valid_o but triggers keccak_run_o and
wait the keccak complete in StMessageWait state. Then after return from
StMessageWait state, it proceed to StPad and send following
keccak_valid_o.

The assumption is revised to check end_of_block and process_latched

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>