[aes] Add initial version of PRNG for cipher core masking
This commit adds an initial version of a high-bandwidth pseudo-random
number generator (PRNG) for masking the AES cipher core. The PRNG
generates pseudo-random data for both the cipher data path and the key
expand module.

Internally the PRNG uses separate LFSRs that are individually
scrambled to break linear shift patterns and muxed onto the output in
alternating fashion. This helps to avoid dependency between successive
sub words (input and output masks).

In addition, the integration of the previously existing low-bandwidth
PRNG for clearing is reworked to better align the two PRNGs. The seeds
of both PRNGs are now exposed as parameters in the top level.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>