[otbn] Add multi-cycle instructions to ISS (LW, BN.LID take 2)
This is reasonably generic, and appears in the documentation with a
note saying "This instruction takes 2 cycles".

On the ISS side, it's a bit of a pity to have to put the call to
add_stall_cycles in the simulator's step function, rather than have it
be part of insn.execute. If we gain any more "generic" code like this,
we should probably make a wrapper on OTBNInsn that does this sort of
thing before passing control to OTBNInsn.execute(). (Much cleaner than
calling super().execute() in every instruction's implementation)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>