[sw/vendor] Move OT-specific code out of riscv_compliance
This change moves OT-specific initialization out of the riscv_compliance
directory, and adds comments explaining why this should be maintained.

In particular, this change makes it possible to change the names of
OpenTitan headers without having to go and generate new patches for
riscv_compliance, and makes the existing Meson target in
riscv_compliance_support somewhat less of a spooky-action-at-a-distance
vector.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>