[util] Instantiate DIF Checklist Too
The SW Checklist was also originally created as a Mako template. This
commit ensures it will be instantated by the `make_new_dif.py` script in
addition to the header template.

Given these changes, it seemed prudent to move the docs into the script,
and document the script in `sw/lib/device/dif/README.md`.

This introduces the `--only` flag which can be used to restrict the
templates that are instantiated, and removes the `--template` flag to
specify a singular template path.

Breaking Change: The `--output` flag now chooses a directory for the
output files, rather than a single file location.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>