[DV common] Add DV_ASSERT_CTRL macro
This commit adds `` `DV_ASSERT_CTRL`` macro that can be used to control
assertions in the DUT from any arbitrary UVM code, without needing hooks
or interfaces between the UVM and the testebench. This uses the
`uvm_config_db` mechanism to allow a configuration bit to be set by the
UVM world to indicate specific instants in the test to turn on / off
assertions in a particular hierarchy.

It also fixes #3623 partially, by avoiding invoking `$assertoff` without
specifying the DUT hierarchy.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>