[DV, TLUL agent] Enable faster B2B requests
Partly addresses the problem described in #3303.
This is an alternate solution to #3417 developed by @rasmus-madsen.

This essentially ensures the when host agent randomizes the TLUL
request, it does not duplicate the `a_source` value within a given
window. That window is the number of outstanding requests a TLUL device
can handle.

An entity that creates and randomizes the TLUL req (sequences, reg
adapter etc) can invoke a function provided in `tl_agent_cfg` to further
constrain the `a_source` to meet the above requirement. This PR provides
the infrastructure to facilitate that.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>