[otbn] Add proper error tracking to otbn_core_model
This turned out to be a bit painful: if you want to update more than
one SystemVerilog variable, you need to put the logic in a task, but
then verilator notices that you're being called from an always_ff
block and tells you to use non-blocking assignments (not
unreasonably). It also complains if you mix blocking and non-blocking
assignments. If you still need to inspect the return code from DPI
functions, you might try to change this pattern:

   retcode = call_dpi(arg1, arg2);
   unique case (retcode)
     ...
   endcase

to this one:

   unique case (call_dpi(arg1, arg2))
     ...
   endcase

but that doesn't work because call_dpi(..) gets evaluated (with
side-effects) for each branch of the case statement until a match is
found.

So I gave up, packing the status bits that I cared about into an
integer and just doing the stateful updates in the C++ code.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>