[kmac] Implement STATUS register
Previously, STATUS register that indicates SHA3 engine was tied to 0.
This commit exposes the sha3 core FSM value. It allows the KMAC top
design to compare the st to check if it is in Idle, Absorbing, or
Squeezing stage.

To support, the definition of SHA3 FSM has been moved to kmac_pkg from
sha3core.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>